将xml格式的网页信息通过xml加载并载入网页文件中待后续处理，生成初步排序的网页库和网页偏移库

+++ include/simhash/Simhasher.hpp
@@ -80,7 +80,7 @@ namespace simhash
                 jenkins _tmp_hasher;
 
                 hashvalues.resize(keywords.size());
-                for (int32_t i = 0; i < keywords.size(); ++i) {
+                for (size_t i = 0; i < keywords.size(); ++i) {
                     hashvalues[i].first = _tmp_hasher(keywords[i].first.c_str(), keywords[i].first.size(), 0);
                     hashvalues[i].second = keywords[i].second;
                 }
